refactors actions into separate files
