+++ example-advanced/src/state/actions/handles/clearPointedHandle.ts
@@ -1,0 +1,6 @@
+import type { Action } from 'state/constants'
+import { mutables } from 'state/mutables'
+
+export const clearPointedHandle: Action = () => {
+  mutables.pointedHandleId = undefined
+}
