+++ example-advanced/src/state/actions/handles/index.ts
@@ -1,0 +1,3 @@
+export * from './setPointedHandle'
+export * from './clearPointedHandle'
+export * from './translateHandle'
