+++ example-advanced/src/state/actions/handles/setPointedHandle.ts
@@ -1,0 +1,6 @@
+import type { Action } from 'state/constants'
+import { mutables } from 'state/mutables'
+
+export const setPointedHandle: Action = (data, payload) => {
+  mutables.pointedHandleId = payload.target
+}
