+++ example-advanced/src/state/actions/history/addToHistory.ts
@@ -1,0 +1,6 @@
+import type { Action } from 'state/constants'
+import { mutables } from '../../mutables'
+
+export const addToHistory: Action = (data) => {
+  mutables.history.push(data)
+}
