+++ example-advanced/src/state/actions/history/index.ts
@@ -1,0 +1,4 @@
+export * from './addToHistory'
+export * from './redo'
+export * from './restore'
+export * from './undo'
