+++ example-advanced/src/state/actions/history/redo.ts
@@ -1,0 +1,7 @@
+import type { Action } from 'state/constants'
+import { mutables } from '../../mutables'
+
+export const redo: Action = (data) => {
+  const snapshot = mutables.history.redo()
+  Object.assign(data, snapshot)
+}
