+++ example-advanced/src/state/actions/selection/clearBrush.ts
@@ -1,0 +1,5 @@
+import type { Action } from 'state/constants'
+
+export const clearBrush: Action = (data) => {
+  data.pageState.brush = undefined
+}
