+++ example-advanced/src/state/actions/selection/clearHoveredShape.ts
@@ -1,0 +1,5 @@
+import type { Action } from 'state/constants'
+
+export const clearHoveredShape: Action = (data) => {
+  data.pageState.hoveredId = undefined
+}
