+++ example-advanced/src/state/actions/selection/clearPointedShape.ts
@@ -1,0 +1,6 @@
+import type { Action } from 'state/constants'
+import { mutables } from 'state/mutables'
+
+export const clearPointedShape: Action = () => {
+  mutables.pointedShapeId = undefined
+}
