+++ example-advanced/src/state/actions/selection/deselectAllShapes.ts
@@ -1,0 +1,5 @@
+import type { Action } from 'state/constants'
+
+export const deselectAllShapes: Action = (data) => {
+  data.pageState.selectedIds = []
+}
