+++ example-advanced/src/state/actions/selection/setHoveredShapeId.ts
@@ -1,0 +1,6 @@
+import type { TLPointerInfo } from '@tldraw/core'
+import type { Action } from 'state/constants'
+
+export const setHoveredShapeId: Action = (data, payload: TLPointerInfo) => {
+  data.pageState.hoveredId = payload.target
+}
