+++ example-advanced/src/state/actions/shapes/index.ts
@@ -1,0 +1,4 @@
+export * from './createArrowShape'
+export * from './createBoxShape'
+export * from './deleteSelectedShapes'
+export * from './updateBoundShapes'
