+++ example-advanced/src/state/actions/snaps/clearSnapInfo.ts
@@ -1,0 +1,6 @@
+import type { Action } from 'state/constants'
+import { mutables } from 'state/mutables'
+
+export const clearSnapInfo: Action = () => {
+  mutables.snapInfo = undefined
+}
