+++ example-advanced/src/state/actions/snaps/clearSnapLines.ts
@@ -1,0 +1,5 @@
+import type { Action } from 'state/constants'
+
+export const clearSnapLines: Action = (data) => {
+  data.overlays.snapLines = []
+}
