+++ example-advanced/src/state/actions/transform/clearPointedBoundsHandle.ts
@@ -1,0 +1,6 @@
+import type { Action } from 'state/constants'
+import { mutables } from 'state/mutables'
+
+export const clearPointedBoundsHandle: Action = (data, payload) => {
+  mutables.pointedBoundsHandleId = undefined
+}
