+++ example-advanced/src/state/actions/transform/index.ts
@@ -1,0 +1,4 @@
+export * from './transformSelectedShapes'
+export * from './setPointedBoundsHandle'
+export * from './setInitialCommonBounds'
+export * from './clearPointedBoundsHandle'
