+++ example-advanced/src/state/actions/transform/setPointedBoundsHandle.ts
@@ -1,0 +1,6 @@
+import type { Action } from 'state/constants'
+import { mutables } from 'state/mutables'
+
+export const setPointedBoundsHandle: Action = (data, payload) => {
+  mutables.pointedBoundsHandleId = payload.target
+}
