+++ example-advanced/src/state/actions/translate/clearIsCloning.ts
@@ -1,0 +1,6 @@
+import type { Action } from 'state/constants'
+import { mutables } from 'state/mutables'
+
+export const clearIsCloning: Action = () => {
+  mutables.isCloning = false
+}
