+++ example-advanced/src/state/actions/translate/index.ts
@@ -1,0 +1,2 @@
+export * from './clearIsCloning'
+export * from './translateSelectedShapes'
